4.2. Результат работы метода [#232622]

+++ src/main/java/ru/job4j/calculator/MathFunc.java
@@ -1,8 +1,8 @@
 package ru.job4j.calculator;
 
 public class MathFunc {
-    public static double func1(double x) {
-        double y = x * x + 1;
+    public static int func1(int x) {
+        int y = x * x + 1;
         return y;
     }
 
@@ -12,9 +12,9 @@ public class MathFunc {
     }
 
     public static void main(String[] args) {
-        double result1 = MathFunc.func1(3);
+        int result1 = MathFunc.func1(3);
         double result2 = MathFunc.func2(5);
-        double result3 = MathFunc.func1(100);
+        int result3 = MathFunc.func1(100);
         double total = result1 + result2;
         System.out.println(total);
         System.out.println(result3);
